+++ src/components/statblock/WeaponAttack.tsx
@@ -7,8 +7,16 @@ function rangedAttack(weapon: Weapon) {
         <i>Hit:</i> {weapon.damage!.average} ({weapon.damage!.diceamount}{weapon.damage!.dice}{weapon.damage!.base ? `+${weapon.damage!.base}` : null}) {weapon.damage!.type} damage.</p>
 }
 
-function meleeAttack(weapon: Weapon) {
-    return <p><i>Melee Weapon Attack:</i> +{weapon.tohit ? weapon.tohit : 5} to hit, one target. <br></br>
+function meleeRange(weapon: Weapon, width: number): number {
+    if (weapon.name.includes("Titan")) {
+        return width/2
+    } else {
+        return 5
+    }
+}
+
+function meleeAttack(weapon: Weapon, width: number) {
+    return <p><i>Melee Weapon Attack:</i> +{weapon.tohit ? weapon.tohit : 5} to hit, reach {meleeRange(weapon, width)}ft. one target. <br></br>
         <i>Hit:</i> {weapon.damage!.average} ({weapon.damage!.diceamount}{weapon.damage!.dice}) {weapon.damage!.type} damage.</p>
 }
 
@@ -16,13 +24,13 @@ function areaAttack(weapon: Weapon) {
     return <p><i>Area Attack:</i> {weapon.area!.size}ft. {weapon.area!.shape}. {weapon.damage!.average} ({weapon.damage!.diceamount}{weapon.damage!.dice}) {weapon.damage!.type} damage. DC{weapon.area!.dc} dexterity halved.</p>
 }
 
-function FormatAttack({ weapon }: { weapon: Weapon }) {
+function FormatAttack({ weapon, width }: { weapon: Weapon, width: number }) {
     if (weapon.damage) {
         return (
             <div>{weapon.range ?
                 rangedAttack(weapon)
                 :
-                (weapon.area ? areaAttack(weapon) : meleeAttack(weapon))
+                (weapon.area ? areaAttack(weapon) : meleeAttack(weapon, width))
             }
             </div>
         )
@@ -32,7 +40,7 @@ function FormatAttack({ weapon }: { weapon: Weapon }) {
 }
 
 
-export default function WeaponAttack({ loadout }: { loadout: Loadout }) {
+export default function WeaponAttack({ loadout, width }: { loadout: Loadout, width: number }) {
 
     return (
         <div>
@@ -40,7 +48,7 @@ export default function WeaponAttack({ loadout }: { loadout: Loadout }) {
                 loadout.getWeapons().map((w: Weapon) =>
                     <div className="property-block">
                         <h4>{w.name}.</h4>
-                        <FormatAttack weapon={w} />
+                        <FormatAttack weapon={w} width={width}/>
                     </div>
                 )
             }